perf: remove per-packet boxing and buffer churn from the update path
Two allocation traces of a playerbots Alterac Valley showed
SMSG_COMPRESSED_UPDATE_OBJECT handling at 62% of all proxy allocation, and
about 44% of everything the proxy allocated was boxed values:

- GetUpdateField<T> (21% of all allocation) ran field.ToString() on every
  call, boxing the enum and building its name for a string-keyed lookup,
  for every field checked on every block. It now reads a per-enum
  FrozenDictionary built once from the same name table through the same
  ToString(), so aliased and undefined values resolve exactly as before.
- HasAnyFlag(IConvertible, IConvertible) boxed both arguments and, for an
  enum, the underlying value again inside ToUInt64: four allocations per
  check, nine checks per object in ComputeValuesChangedMask, three passes
  per object (14%). Same-type generic and primitive overloads replace it,
  and it is deleted outright, so a check that mixes types now fails to
  compile instead of quietly boxing. The 63 mixed call sites cast the enum
  constant to the field's type, which cannot change the result because the
  field's upper bits are zero.

Buffer and object churn on the same path:

- ServerPacket creates its WorldPacket on first write, and ByteBuffer rents
  on first write. ISpanWritable packets and unsent aura/power scratch
  packets no longer hold a rental, so the pool stops allocating a fresh
  array for each of them.
- The receive loop's LegacyServerPacketHeader is a struct, and
  ReceiveBufferFully uses a pooled ValueTask builder.
- Packed GUIDs are built on the stack instead of a byte[8] per half.
- One pair of update masks per client is refilled for every Values block.
- A player update only allocates its quest log when a quest slot is
  present.
- FilterV3_4_3Values compacts in place instead of a per-call RemoveAll
  closure; its summary and the create-registered trace no longer build
  strings with logging off.
- The speed-change handlers use precomputed opcode tables instead of
  renaming the opcode through strings on every packet.
- MonsterMove point lists are sized up front.
- UpdateObject.Write disposes its scratch buffers and splices them in as
  spans instead of copying out through GetData twice.
- The metrics summary sorts in a pooled buffer.

Measured on the Mac mini against verbatim copies of the old code
(AllocationHotSpotBenchmarks): 30 field lookups 381 ns/725 B -> 15 ns/0 B,
nine flag checks 120 ns/864 B -> 2.3 ns/0 B, 16 packed GUIDs
322 ns/1024 B -> 179 ns/0 B, constructing an unsent packet
252 ns/182 B with Gen1/Gen2 -> 7 ns/88 B.

Live, same Alterac Valley setup, busy minutes before -> after:
35-66 KB -> 7-13 KB allocated per compressed update, handler average
0.17-0.24 ms -> 0.11-0.13 ms, p95 0.32-0.55 ms -> 0.19-0.24 ms, gen0
collections 20-27/min -> 5-10/min, gen1 3-13/min -> 0-2/min.

+++ HermesProxy/World/Client/PacketHandlers/CombatHandler.cs
@@ -105,11 +105,11 @@ public partial class WorldClient
             subDmg.IntDamage = packet.ReadInt32();
 
             if (LegacyVersion.RemovedInVersion(ClientVersionBuild.V3_0_3_9183) ||
-                hitInfo.HasAnyFlag(HitInfo.PartialAbsorb | HitInfo.FullAbsorb))
+                hitInfo.HasAnyFlag((uint)(HitInfo.PartialAbsorb | HitInfo.FullAbsorb)))
                 subDmg.Absorbed = packet.ReadInt32();
 
             if (LegacyVersion.RemovedInVersion(ClientVersionBuild.V3_0_3_9183) ||
-                hitInfo.HasAnyFlag(HitInfo.PartialResist | HitInfo.FullResist))
+                hitInfo.HasAnyFlag((uint)(HitInfo.PartialResist | HitInfo.FullResist)))
                 subDmg.Resisted = packet.ReadInt32();
 
             attack.SubDmg.Add(subDmg);
@@ -124,13 +124,13 @@ public partial class WorldClient
         attack.MeleeSpellID = packet.ReadUInt32();
 
         if (LegacyVersion.RemovedInVersion(ClientVersionBuild.V3_0_3_9183) ||
-            hitInfo.HasAnyFlag(HitInfo.Block))
+            hitInfo.HasAnyFlag((uint)HitInfo.Block))
             attack.BlockAmount = packet.ReadInt32();
 
-        if (hitInfo.HasAnyFlag(HitInfo.RageGain))
+        if (hitInfo.HasAnyFlag((uint)HitInfo.RageGain))
             attack.RageGained = packet.ReadInt32();
 
-        if (hitInfo.HasAnyFlag(HitInfo.Unk0))
+        if (hitInfo.HasAnyFlag((uint)HitInfo.Unk0))
         {
             attack.UnkState = new();
             attack.UnkState.State1 = packet.ReadUInt32();
